마이페이지 투두 통계 관련 패키지를 분리하고 url을 변경하라
